+++ internal/processor/remove_user.go
@@ -75,7 +75,7 @@ func (p *processor) handleRemoveUserAction(msg data.ModulePayload) error {
 			}
 
 			if dbUser.Id == nil {
-				err = p.sendDeleteUser(msg.RequestId, *dbUser)
+				err = p.SendDeleteUser(msg.RequestId, *dbUser)
 				if err != nil {
 					p.log.WithError(err).Errorf("failed to publish delete user for message action with id `%s`", msg.RequestId)
 					return errors.Wrap(err, "failed to publish delete user")