+++ internal/worker/main.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"time"
 
+	"github.com/google/uuid"
 	"gitlab.com/distributed_lab/acs/telegram-module/internal/config"
 	"gitlab.com/distributed_lab/acs/telegram-module/internal/data"
 	"gitlab.com/distributed_lab/acs/telegram-module/internal/data/postgres"
@@ -20,16 +21,20 @@ type Worker interface {
 }
 
 type worker struct {
-	logger    *logan.Entry
-	processor processor.Processor
-	linksQ    data.Links
+	logger       *logan.Entry
+	processor    processor.Processor
+	linksQ       data.Links
+	permissionsQ data.Permissions
+	usersQ       data.Users
 }
 
 func NewWorker(cfg config.Config, ctx context.Context) Worker {
 	return &worker{
-		logger:    cfg.Log().WithField("runner", serviceName),
-		processor: processor.NewProcessor(cfg, ctx),
-		linksQ:    postgres.NewLinksQ(cfg.DB()),
+		logger:       cfg.Log().WithField("runner", serviceName),
+		processor:    processor.NewProcessor(cfg, ctx),
+		linksQ:       postgres.NewLinksQ(cfg.DB()),
+		permissionsQ: postgres.NewPermissionsQ(cfg.DB()),
+		usersQ:       postgres.NewUsersQ(cfg.DB()),
 	}
 }
 
@@ -48,6 +53,8 @@ func (w *worker) Run(ctx context.Context) {
 func (w *worker) processPermissions(_ context.Context) error {
 	w.logger.Info("fetching links")
 
+	startTime := time.Now()
+
 	links, err := w.linksQ.Select()
 	if err != nil {
 		return errors.Wrap(err, "failed to get links")
@@ -58,7 +65,7 @@ func (w *worker) processPermissions(_ context.Context) error {
 		w.logger.Info("no links were found")
 		return nil
 	}
-
+	
 	w.logger.Infof("found %v links", reqAmount)
 
 	for _, link := range links {
@@ -73,6 +80,72 @@ func (w *worker) processPermissions(_ context.Context) error {
 		w.logger.Infof("successfully processed link `%s`", link.Link)
 	}
 
+	err = w.removeOldUsers(startTime)
+	if err != nil {
+		w.logger.WithError(err).Errorf("failed to remove old users")
+		return errors.Wrap(err, "failed to remove old users")
+	}
+
+	err = w.removeOldPermissions(startTime)
+	if err != nil {
+		w.logger.WithError(err).Errorf("failed to remove old permissions")
+		return errors.Wrap(err, "failed to remove old permissions")
+	}
+
+	return nil
+}
+
+func (w *worker) removeOldUsers(borderTime time.Time) error {
+	w.logger.Infof("started removing old users")
+
+	users, err := w.usersQ.FilterByLowerTime(borderTime).Select()
+	if err != nil {
+		w.logger.Infof("failed to select users")
+		return errors.Wrap(err, " failed to select users")
+	}
+
+	w.logger.Infof("found `%d` users to delete", len(users))
+
+	for _, user := range users {
+		if user.Id == nil { //if unverified user we need to remove them from `unverified-svc`
+			err = w.processor.SendDeleteUser(uuid.New().String(), user)
+			if err != nil {
+				w.logger.WithError(err).Errorf("failed to publish delete user")
+				return errors.Wrap(err, " failed to publish delete user")
+			}
+		}
+
+		err = w.usersQ.Delete(user.TelegramId)
+		if err != nil {
+			w.logger.Infof("failed to delete user with telegram id `%d`", user.TelegramId)
+			return errors.Wrap(err, " failed to delete user")
+		}
+	}
+
+	w.logger.Infof("finished removing old users")
+	return nil
+}
+
+func (w *worker) removeOldPermissions(borderTime time.Time) error {
+	w.logger.Infof("started removing old permissions")
+
+	permissions, err := w.permissionsQ.FilterByLowerTime(borderTime).Select()
+	if err != nil {
+		w.logger.Infof("failed to select permissions")
+		return errors.Wrap(err, " failed to select permissions")
+	}
+
+	w.logger.Infof("found `%d` permissions to delete", len(permissions))
+
+	for _, permission := range permissions {
+		err = w.permissionsQ.Delete(permission.TelegramId, permission.Link)
+		if err != nil {
+			w.logger.Infof("failed to delete permission")
+			return errors.Wrap(err, " failed to delete permission")
+		}
+	}
+
+	w.logger.Infof("finished removing old permissions")
 	return nil
 }