+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2023-03-27
+
+### Added
+
+- Worker delete users and permissions that wasn't updated for some time
+- `updated_at` column in users table
+
 ## [1.0.3] - 2023-03-24
 
 ### Added